@@ -2,47 +2,56 @@
   "name": "dr-kobros",
   "version": "0.1.0",
   "private": true,
+  "type": "module",
   "scripts": {
-    "build": "next build",
-    "dev": "next dev",
-    "lint": "next lint",
+    "build": "next build --webpack",
+    "dev": "next dev --webpack",
+    "lint": "eslint",
     "start": "next start"
   },
   "dependencies": {
-    "axios": "^1.8.4",
-    "chart.js": "^4.4.9",
+    "axios": "^1.13.1",
+    "chart.js": "^4.5.1",
     "clsx": "^2.1.1",
-    "luxon": "^3.6.1",
-    "mongodb": "^6.15.0",
-    "next": "15.2.4",
+    "luxon": "^3.7.2",
+    "mongodb": "^6.20.0",
+    "next": "16.0.1",
     "normalize.css": "^8.0.1",
-    "ramda": "^0.30.1",
-    "react": "^19.1.0",
-    "react-chartjs-2": "^5.3.0",
-    "react-dom": "^19.1.0",
+    "openai": "^6.7.0",
+    "ramda": "^0.32.0",
+    "react": "^19.2.0",
+    "react-chartjs-2": "^5.3.1",
+    "react-dom": "^19.2.0",
     "react-icons": "^5.5.0",
-    "sharp": "^0.34.1",
-    "uuid": "^11.1.0",
-    "yup": "^1.6.1"
+    "react-markdown": "^10.1.0",
+    "sharp": "^0.34.4",
+    "uuid": "^13.0.0",
+    "workflow": "4.0.1-beta.6",
+    "yup": "^1.7.1"
   },
   "devDependencies": {
-    "@next/bundle-analyzer": "^15.3.1",
-    "@types/luxon": "^3.6.2",
-    "@types/node": "^22.14.1",
-    "@types/ramda": "^0.30.2",
-    "@types/react": "^19.1.2",
-    "@types/react-dom": "^19.1.2",
-    "@types/uuid": "^10.0.0",
-    "@typescript-eslint/eslint-plugin": "^8.30.1",
-    "@vanilla-extract/css": "^1.17.1",
-    "@vanilla-extract/next-plugin": "^2.4.10",
-    "eslint": "^9.25.0",
-    "eslint-config-next": "^15.3.1",
-    "eslint-config-prettier": "^10.1.2",
-    "eslint-plugin-prettier": "^5.2.6",
-    "prettier": "^3.5.3",
-    "typescript": "^5.8.3",
-    "typescript-eslint": "^8.30.1"
+    "@eslint/js": "^9.38.0",
+    "@next/bundle-analyzer": "^16.0.1",
+    "@types/luxon": "^3.7.1",
+    "@types/node": "^24.9.2",
+    "@types/ramda": "^0.31.1",
+    "@types/react": "^19.2.2",
+    "@types/react-dom": "^19.2.2",
+    "@typescript-eslint/eslint-plugin": "^8.46.2",
+    "@vanilla-extract/css": "^1.17.4",
+    "@vanilla-extract/next-plugin": "^2.4.14",
+    "babel-plugin-react-compiler": "^1.0.0",
+    "eslint": "^9.38.0",
+    "eslint-config-next": "^16.0.1",
+    "eslint-config-prettier": "^10.1.8",
+    "eslint-plugin-jsx-a11y": "^6.10.2",
+    "eslint-plugin-prettier": "^5.5.4",
+    "eslint-plugin-react": "^7.37.5",
+    "eslint-plugin-react-hooks": "^7.0.1",
+    "prettier": "^3.6.2",
+    "tsx": "^4.20.6",
+    "typescript": "^5.9.3",
+    "typescript-eslint": "^8.46.2"
   },
-  "packageManager": "pnpm@10.8.1+sha512.c50088ba998c67b8ca8c99df8a5e02fd2ae2e2b29aaf238feaa9e124248d3f48f9fb6db2424949ff901cffbb5e0f0cc1ad6aedb602cd29450751d11c35023677"
+  "packageManager": "pnpm@10.20.0"
 }
